+++ lib/QuizScreens/question2.dart
@@ -1,30 +1,28 @@
 import 'package:flutter/material.dart';
-import 'package:book_rent_app_chapter3/QuizScreens/Question9.dart';
-import 'package:book_rent_app_chapter3/model/QuizDataModel.dart';
+import 'package:book_rent_app_chapter3/QuizScreens/question3.dart';
+import 'package:book_rent_app_chapter3/model/quiz_data_model.dart';
 
 //ignore: must_be_immutable
-class Question8 extends StatefulWidget {
+class Question2 extends StatefulWidget {
   QuizData quizData = QuizData();
-  Question8({Key? key, required this.quizData}) : super(key: key);
+  Question2({Key? key, required this.quizData}) : super(key: key);
 
   @override
-  _Question8State createState() => _Question8State();
+  _Question2State createState() => _Question2State();
 
-  int csms = 0;
-  int eu = 0;
-  int pw = 0;
-  int fsg = 0;
-  int kr = 0;
+  int hppa = 0;
   int tug = 0;
+  int pfo = 0;
+  int tgdt = 0;
 }
 
-class _Question8State extends State<Question8> {
+class _Question2State extends State<Question2> {
   String? chosen;
   bool answered = false;
-  String option1 = 'I wanna be happy and filled with\npositivity';
-  String option2 = 'Tear jerker, Hit me with a solid\nemotional story';
-  String option3 = 'Hot and steamy romance pls';
-  String option4 = 'Send my brain for a jog, give me facts\nand knowledge';
+  String option1 = 'Fancy restaurant';
+  String option2 = 'Backpacking through Europe';
+  String option3 = 'An adventurous date in a fantasy\nworld';
+  String option4 = 'Pizza date with fav true crime\npodcasts';
 
   @override
   Widget build(BuildContext context) {
@@ -42,7 +40,7 @@ class _Question8State extends State<Question8> {
                       context,
                       MaterialPageRoute(
                           builder: (c) =>
-                              Question9(
+                              Question3(
                                 quizData: widget.quizData,
                               )));
                 },
@@ -54,7 +52,7 @@ class _Question8State extends State<Question8> {
               Row(
                 children: const [
                   Text(
-                    'Question 8',
+                    'Question 2',
                     style: TextStyle(fontSize: 20, color: Colors.black),
                   )
                 ],
@@ -65,7 +63,7 @@ class _Question8State extends State<Question8> {
               Row(
                 children: const [
                   Text(
-                    'What is your favorite part about a book?',
+                    'Where would you like to go on a date?',
                     style: TextStyle(fontSize: 20, color: Colors.black),
                   ),
                 ],
@@ -102,7 +100,8 @@ class _Question8State extends State<Question8> {
                             Text(
                               option1,
                               style:
-                              const TextStyle(color: Colors.black, fontSize: 20),
+                              const TextStyle(color: Colors.black,
+                                  fontSize: 20),
                             )
                           ]))),
                   GestureDetector(
@@ -111,8 +110,8 @@ class _Question8State extends State<Question8> {
                           setState(() {
                             chosen = option2;
                             answered = true;
-                            widget.kr += 1;
-                            widget.quizData.kr += widget.kr;
+                            widget.pfo += 1;
+                            widget.quizData.pfo += widget.pfo;
                           });
                         }
                       },
@@ -132,7 +131,8 @@ class _Question8State extends State<Question8> {
                             Text(
                               option2,
                               style:
-                              const TextStyle(color: Colors.black, fontSize: 20),
+                              const TextStyle(color: Colors.black,
+                                  fontSize: 20),
                             )
                           ]))),
                   GestureDetector(
@@ -141,8 +141,8 @@ class _Question8State extends State<Question8> {
                           setState(() {
                             chosen = option3;
                             answered = true;
-                            widget.fsg += 1;
-                            widget.quizData.fsg += widget.fsg;
+                            widget.hppa += 1;
+                            widget.quizData.hppa += widget.hppa;
                           });
                         }
                       },
@@ -162,7 +162,8 @@ class _Question8State extends State<Question8> {
                             Text(
                               option3,
                               style:
-                              const TextStyle(color: Colors.black, fontSize: 20),
+                              const TextStyle(color: Colors.black,
+                                  fontSize: 20),
                             )
                           ]))),
                   GestureDetector(
@@ -171,12 +172,8 @@ class _Question8State extends State<Question8> {
                           setState(() {
                             chosen = option4;
                             answered = true;
-                            widget.csms += 1;
-                            widget.eu += 1;
-                            widget.pw += 1;
-                            widget.quizData.csms += widget.csms;
-                            widget.quizData.eu += widget.eu;
-                            widget.quizData.pw += widget.pw;
+                            widget.tgdt += 1;
+                            widget.quizData.tgdt += widget.tgdt;
                           });
                         }
                       },
@@ -196,7 +193,8 @@ class _Question8State extends State<Question8> {
                             Text(
                               option4,
                               style:
-                              const TextStyle(color: Colors.black, fontSize: 20),
+                              const TextStyle(color: Colors.black,
+                                  fontSize: 20),
                             )
                           ])))
                 ],
@@ -210,22 +208,18 @@ class _Question8State extends State<Question8> {
                     if (chosen == option1) {
                       widget.quizData.tug -= 1;
                     } else if (chosen == option2) {
-                      widget.quizData.kr -= 1;
+                      widget.quizData.pfo -= 1;
                     } else if (chosen == option3) {
-                      widget.quizData.fsg -= 1;
+                      widget.quizData.hppa -= 1;
                     } else if (chosen == option4) {
-                      widget.quizData.eu -= 1;
-                      widget.quizData.csms -= 1;
-                      widget.quizData.pw -= 1;
+                      widget.quizData.tgdt -= 1;
                     }
                     chosen = '';
                     answered = false;
+                    widget.tgdt = 0;
+                    widget.pfo = 0;
+                    widget.hppa = 0;
                     widget.tug = 0;
-                    widget.fsg = 0;
-                    widget.kr = 0;
-                    widget.eu = 0;
-                    widget.csms = 0;
-                    widget.pw = 0;
                   });
                 },
                 child: const Text('Reset'),
