+++ lib/QuizScreens/question8.dart
@@ -1,28 +1,30 @@
 import 'package:flutter/material.dart';
-import 'package:book_rent_app_chapter3/QuizScreens/Question3.dart';
-import 'package:book_rent_app_chapter3/model/QuizDataModel.dart';
+import 'package:book_rent_app_chapter3/QuizScreens/question9.dart';
+import 'package:book_rent_app_chapter3/model/quiz_data_model.dart';
 
 //ignore: must_be_immutable
-class Question2 extends StatefulWidget {
+class Question8 extends StatefulWidget {
   QuizData quizData = QuizData();
-  Question2({Key? key, required this.quizData}) : super(key: key);
+  Question8({Key? key, required this.quizData}) : super(key: key);
 
   @override
-  _Question2State createState() => _Question2State();
+  _Question8State createState() => _Question8State();
 
-  int hppa = 0;
+  int csms = 0;
+  int eu = 0;
+  int pw = 0;
+  int fsg = 0;
+  int kr = 0;
   int tug = 0;
-  int pfo = 0;
-  int tgdt = 0;
 }
 
-class _Question2State extends State<Question2> {
+class _Question8State extends State<Question8> {
   String? chosen;
   bool answered = false;
-  String option1 = 'Fancy restaurant';
-  String option2 = 'Backpacking through Europe';
-  String option3 = 'An adventurous date in a fantasy\nworld';
-  String option4 = 'Pizza date with fav true crime\npodcasts';
+  String option1 = 'I wanna be happy and filled with\npositivity';
+  String option2 = 'Tear jerker, Hit me with a solid\nemotional story';
+  String option3 = 'Hot and steamy romance pls';
+  String option4 = 'Send my brain for a jog, give me facts\nand knowledge';
 
   @override
   Widget build(BuildContext context) {
@@ -40,7 +42,7 @@ class _Question2State extends State<Question2> {
                       context,
                       MaterialPageRoute(
                           builder: (c) =>
-                              Question3(
+                              Question9(
                                 quizData: widget.quizData,
                               )));
                 },
@@ -52,7 +54,7 @@ class _Question2State extends State<Question2> {
               Row(
                 children: const [
                   Text(
-                    'Question 2',
+                    'Question 8',
                     style: TextStyle(fontSize: 20, color: Colors.black),
                   )
                 ],
@@ -63,7 +65,7 @@ class _Question2State extends State<Question2> {
               Row(
                 children: const [
                   Text(
-                    'Where would you like to go on a date?',
+                    'What is your favorite part about a book?',
                     style: TextStyle(fontSize: 20, color: Colors.black),
                   ),
                 ],
@@ -100,8 +102,7 @@ class _Question2State extends State<Question2> {
                             Text(
                               option1,
                               style:
-                              const TextStyle(color: Colors.black,
-                                  fontSize: 20),
+                              const TextStyle(color: Colors.black, fontSize: 20),
                             )
                           ]))),
                   GestureDetector(
@@ -110,8 +111,8 @@ class _Question2State extends State<Question2> {
                           setState(() {
                             chosen = option2;
                             answered = true;
-                            widget.pfo += 1;
-                            widget.quizData.pfo += widget.pfo;
+                            widget.kr += 1;
+                            widget.quizData.kr += widget.kr;
                           });
                         }
                       },
@@ -131,8 +132,7 @@ class _Question2State extends State<Question2> {
                             Text(
                               option2,
                               style:
-                              const TextStyle(color: Colors.black,
-                                  fontSize: 20),
+                              const TextStyle(color: Colors.black, fontSize: 20),
                             )
                           ]))),
                   GestureDetector(
@@ -141,8 +141,8 @@ class _Question2State extends State<Question2> {
                           setState(() {
                             chosen = option3;
                             answered = true;
-                            widget.hppa += 1;
-                            widget.quizData.hppa += widget.hppa;
+                            widget.fsg += 1;
+                            widget.quizData.fsg += widget.fsg;
                           });
                         }
                       },
@@ -162,8 +162,7 @@ class _Question2State extends State<Question2> {
                             Text(
                               option3,
                               style:
-                              const TextStyle(color: Colors.black,
-                                  fontSize: 20),
+                              const TextStyle(color: Colors.black, fontSize: 20),
                             )
                           ]))),
                   GestureDetector(
@@ -172,8 +171,12 @@ class _Question2State extends State<Question2> {
                           setState(() {
                             chosen = option4;
                             answered = true;
-                            widget.tgdt += 1;
-                            widget.quizData.tgdt += widget.tgdt;
+                            widget.csms += 1;
+                            widget.eu += 1;
+                            widget.pw += 1;
+                            widget.quizData.csms += widget.csms;
+                            widget.quizData.eu += widget.eu;
+                            widget.quizData.pw += widget.pw;
                           });
                         }
                       },
@@ -193,8 +196,7 @@ class _Question2State extends State<Question2> {
                             Text(
                               option4,
                               style:
-                              const TextStyle(color: Colors.black,
-                                  fontSize: 20),
+                              const TextStyle(color: Colors.black, fontSize: 20),
                             )
                           ])))
                 ],
@@ -208,18 +210,22 @@ class _Question2State extends State<Question2> {
                     if (chosen == option1) {
                       widget.quizData.tug -= 1;
                     } else if (chosen == option2) {
-                      widget.quizData.pfo -= 1;
+                      widget.quizData.kr -= 1;
                     } else if (chosen == option3) {
-                      widget.quizData.hppa -= 1;
+                      widget.quizData.fsg -= 1;
                     } else if (chosen == option4) {
-                      widget.quizData.tgdt -= 1;
+                      widget.quizData.eu -= 1;
+                      widget.quizData.csms -= 1;
+                      widget.quizData.pw -= 1;
                     }
                     chosen = '';
                     answered = false;
-                    widget.tgdt = 0;
-                    widget.pfo = 0;
-                    widget.hppa = 0;
                     widget.tug = 0;
+                    widget.fsg = 0;
+                    widget.kr = 0;
+                    widget.eu = 0;
+                    widget.csms = 0;
+                    widget.pw = 0;
                   });
                 },
                 child: const Text('Reset'),
