+++ lib/QuizScreens/quiz_result.dart
@@ -1,7 +1,7 @@
 // ignore_for_file: avoid_print
 
 import 'package:flutter/material.dart';
-import 'package:book_rent_app_chapter3/model/QuizDataModel.dart';
+import 'package:book_rent_app_chapter3/model/quiz_data_model.dart';
 import 'package:quiver/iterables.dart' as quiver;
 
 //ignore: must_be_immutable
