+++ lib/Screens/all_books_of_a_genre.dart
@@ -1,25 +1,29 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:book_rent_app_chapter3/provider/product.dart';
-import 'package:book_rent_app_chapter3/Widgets.dart';
+import 'package:book_rent_app_chapter3/widgets.dart';
 
-class Recommendations extends StatefulWidget {
+class AllBooksofAGenre extends StatefulWidget {
+  final String genre;
+
+  const AllBooksofAGenre({Key? key,required this.genre}) : super(key: key);
+
   @override
-  _RecommendationsState createState() => _RecommendationsState();
+  _AllBooksofAGenreState createState() => _AllBooksofAGenreState();
 }
 
-class _RecommendationsState extends State<Recommendations> {
+class _AllBooksofAGenreState extends State<AllBooksofAGenre> {
   @override
   Widget build(BuildContext context) {
     final productProvider = Provider.of<ProductProvider>(context);
     return SafeArea(
       child: Container(
         color: Colors.pink[50],
-        child: ListView(
+        child: Column(
           children: productProvider.products
               .map((item) =>
               GestureDetector(
-                child: item.recommended
+                child: item.genre == widget.genre
                     ? ProductCard(
                   product: item,
                 )
