+++ lib/Screens/best_sellers.dart
@@ -1,29 +1,25 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:book_rent_app_chapter3/provider/product.dart';
-import 'package:book_rent_app_chapter3/Widgets.dart';
+import 'package:book_rent_app_chapter3/widgets.dart';
 
-class AllBooksofAGenre extends StatefulWidget {
-  final String genre;
-
-  const AllBooksofAGenre({Key? key,required this.genre}) : super(key: key);
-
+class BestSellers extends StatefulWidget {
   @override
-  _AllBooksofAGenreState createState() => _AllBooksofAGenreState();
+  _BestSellersState createState() => _BestSellersState();
 }
 
-class _AllBooksofAGenreState extends State<AllBooksofAGenre> {
+class _BestSellersState extends State<BestSellers> {
   @override
   Widget build(BuildContext context) {
     final productProvider = Provider.of<ProductProvider>(context);
     return SafeArea(
       child: Container(
         color: Colors.pink[50],
-        child: Column(
+        child: ListView(
           children: productProvider.products
               .map((item) =>
               GestureDetector(
-                child: item.genre == widget.genre
+                child: item.bestseller
                     ? ProductCard(
                   product: item,
                 )
