+++ lib/Screens/login_page.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
-import 'package:book_rent_app_chapter3/Screens/Loading.dart';
-import 'package:book_rent_app_chapter3/Screens/Register.dart';
+import 'package:book_rent_app_chapter3/Screens/loading.dart';
+import 'package:book_rent_app_chapter3/Screens/register.dart';
 
 import 'package:book_rent_app_chapter3/provider/user.dart';
