+++ lib/Screens/product_details.dart
@@ -2,7 +2,7 @@ import 'package:carousel_pro_nullsafety/carousel_pro_nullsafety.dart';
 
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:book_rent_app_chapter3/Screens/Loading.dart';
+import 'package:book_rent_app_chapter3/Screens/loading.dart';
 
 import 'package:book_rent_app_chapter3/model/product.dart';
 import 'package:book_rent_app_chapter3/provider/app.dart';
