+++ lib/main.dart
@@ -4,8 +4,8 @@ import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:book_rent_app_chapter3/Screens/LoginPage.dart';
-import 'package:book_rent_app_chapter3/Screens/NavBar.dart';
+import 'package:book_rent_app_chapter3/Screens/login_page.dart';
+import 'package:book_rent_app_chapter3/Screens/nav_bar.dart';
 import 'package:book_rent_app_chapter3/Screens/splash.dart';
 import 'package:book_rent_app_chapter3/provider/app.dart';
 import 'package:book_rent_app_chapter3/provider/product.dart';
